+++ lib/messages.js
@@ -16,6 +16,10 @@ function addMessage(opts) {
     }
 
 
+    if (!utils.isOutgoingAllowed(doc)) {
+        opts.state = 'denied';
+    }
+
     if (phone && opts.message) {
         try {
             utils.addMessage(doc, {
