+++ lib/utils.js
@@ -1,6 +1,7 @@
 var _ = require('underscore'),
     uuid = require('uuid'),
     moment = require('moment'),
+    libphonenumber = require('libphonenumber'),
     template = require('./template'),
     config = require('../config'),
     i18n = require('../i18n'),
@@ -362,7 +363,8 @@ module.exports = {
             _.first(msg.messages).message = options.message;
     },
     addScheduledMessage: function(doc, options) {
-        var options = options || {},
+        var self = module.exports,
+            options = options || {},
             due = options.due,
             message = options.message,
             phone = applyPhoneFilters(config, options.phone);
@@ -384,7 +386,13 @@ module.exports = {
                 message: message
             }]
         };
-        setTaskState(task, doc.muted ? 'muted' : 'scheduled');
+
+        if (!self.isOutgoingAllowed(doc)) {
+            setTaskState(task, 'denied');
+        } else {
+            setTaskState(task, 'scheduled');
+        }
+
         _.extend(task, options);
         doc.scheduled_tasks.push(task);
     },
@@ -537,5 +545,47 @@ module.exports = {
     },
     escapeRegex: function(s) {
         return s.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
+    },
+    /*
+     * Return false when the recipient phone matches the denied list.
+     *
+     * outgoing_deny_list is a comma separated list of strings. If a string in
+     * that list matches the beginning of the phone then we set up a response
+     * with a denied state. The pending message process will ignore these
+     * messages and those reports will be left without an auto-reply. The
+     * denied messages still show up in the messages export.
+     *
+     * @param {String} from - Recipient phone number
+     * @returns {Boolean}
+     */
+    isOutgoingAllowed: function(from) {
+        var self = module.exports,
+            conf = config.get('outgoing_deny_list') || '';
+        if (!from) {
+            return true;
+        }
+        if (self._isMessageFromGateway(from)) {
+            return false;
+        }
+        return _.every(conf.split(','), function(s) {
+            // ignore falsey inputs
+            if (!s) {
+                return true;
+            }
+            return !from.match(
+                new RegExp('^' + self.escapeRegex(s.trim()), 'i')
+            );
+        });
+    },
+    /*
+     * Used to avoid infinite loops of auto-reply messages between gateway and
+     * itself.
+     */
+    _isMessageFromGateway: function(from) {
+        var gw = config.get('gateway_number');
+        if (typeof gw === 'string' && typeof from === 'string') {
+            return libphonenumber.phoneUtil.isNumberMatch(gw, from) >= 3;
+        }
+        return false;
     }
 }
