+++ test/unit/utils.js
@@ -6,13 +6,17 @@ var _ = require('underscore'),
     utils = require('../../lib/utils'),
     config = require('../../config');
 
+var restore = function(objs) {
+    _.each(objs, function(obj) {
+        if (obj.restore) obj.restore();
+    });
+}
+
 exports.tearDown = function(callback) {
-    if (db.view.restore) {
-        db.view.restore();
-    }
-    if (config.get.restore) {
-        config.get.restore();
-    }
+    restore([
+        config.get,
+        db.view
+    ]);
     callback();
 }
 
@@ -433,3 +437,56 @@ exports['translate returns key if translations not found'] = function(test) {
     test.equals(utils.translate('sms_received'), 'sms_received');
     test.done();
 }
+
+exports['describe isOutgoingAllowed'] = function(test) {
+    /*
+     * Support comma separated string config to match an outgoing phone number
+     * or MNO (mobile network operator) defined string.
+     */
+    var tests = [
+      // denied
+      ['+123', '+123', false],
+      ['+123', '+123999999', false],
+      ['SAFARI', 'SAFARICOM', false],
+      ['Safari', 'SAFARICOM', false],
+      ['+123,+456,+789', '+456', false],
+      ['+123,+456,+789', '+4569999999', false],
+      ['SAFARI, ORANGE', 'ORANGE NET', false],
+      ['0', '0000123', false],
+      ['0', '0', false],
+      // allowed
+      ['+123', '+999', true],
+      ['SAFARI, ORANGE NET', 'ORANGE', true],
+      ['VIVO', 'EM VIVO', true],
+      ['0', '-1', true],
+      // allow falsey inputs
+      ['snarf', undefined, true],
+      ['snarf', null, true],
+      ['', '+123', true],
+      ['', '', true]
+    ];
+    _.each(tests, function(t) {
+      var s = sinon.stub(config, 'get');
+      s.withArgs('outgoing_deny_list').returns(t[0]);
+      test.equals(utils.isOutgoingAllowed(t[1]), t[2]);
+      s.restore();
+    });
+    test.done();
+};
+
+exports['describe _isMessageFromGateway'] = function(test) {
+    var tests = [
+      ['+774455558888', '77-44-5555-8888', true],
+      ['+774455558889', '77-44-5555-8888', false],
+      // missing country code matches
+      ['+41446681800', '446681800', true]
+    ];
+    _.each(tests, function(t) {
+      var s = sinon.stub(config, 'get');
+      s.withArgs('gateway_number').returns(t[0]);
+      test.equals(utils._isMessageFromGateway(t[1]), t[2]);
+      s.restore();
+    });
+    test.done();
+};
+
