+++ transitions/default_responses.js
@@ -1,6 +1,5 @@
 var _ = require('underscore'),
     moment = require('moment'),
-    libphonenumber = require('libphonenumber'),
     config = require('../config'),
     utils = require('../lib/utils'),
     logger = require('../lib/logger'),
@@ -16,42 +15,6 @@ module.exports = {
             && !doc.kujua_message
             && self._isReportedAfterStartDate(doc)
         );
-    },
-    /*
-     * Avoid infinite loops of auto-reply messages between gateway and itself.
-     */
-    _isMessageFromGateway: function(doc) {
-        var self = module.exports,
-            gw = self._getConfig('gateway_number');
-        if (typeof gw === 'string' && typeof doc.from === 'string') {
-            return libphonenumber.phoneUtil.isNumberMatch(gw, doc.from) >= 3;
-        }
-        return false;
-    },
-    /*
-     * Return false when the recipient phone matches the denied list.
-     *
-     * outgoing_deny_list is a comma separated list of strings. If a string in
-     * that list matches the beginning of the phone then we set up a response
-     * with a denied state. The pending message process will ignore these
-     * messages and those reports will be left without an auto-reply. The
-     * denied messages still show up in the messages export.
-     */
-    _isResponseAllowed: function(doc) {
-        var self = module.exports,
-            conf = self._getConfig('outgoing_deny_list') || '';
-        if (self._isMessageFromGateway(doc)) {
-            return false;
-        }
-        return _.every(conf.split(','), function(s) {
-            // ignore falsey inputs
-            if (!s || !doc.from) {
-                return true;
-            }
-            return !doc.from.match(
-                new RegExp('^' + utils.escapeRegex(s.trim()), 'i')
-            );
-        });
     },
     _isReportedAfterStartDate: function(doc) {
         var self = module.exports,
@@ -107,9 +70,6 @@ module.exports = {
                 phone: doc.from,
                 message: msg
             };
-        if (!self._isResponseAllowed(doc)) {
-            opts.state = 'denied';
-        }
         messages.addMessage(opts);
     },
     onMatch: function(change, db, audit, callback) {
